show full week button (+ refactoring)
