+++ client/src/app/calendar/_components/Timeline.tsx
@@ -3,11 +3,11 @@ import { lazy } from 'react';
 import { theme } from '@/util/tailwindVars';
 import { LayoutGroup } from 'framer-motion';
 
-import { useDatesArray } from '../_util/dateUtils';
+import { useDatesArray } from '../_util/datesArray';
 import { CalendarProps } from './Calendar';
 import { gridCols } from '../_util/grid';
 import { useGetRooms } from '../new/state/getRoomData';
-import { useDisplayByRooms } from '../_util/displayByRooms';
+import { useDisplayByRooms } from '../_util/queryStates';
 
 import TimelineHeader, { TimelineHeaderFrame } from './TimelineHeader';
